# apps/forwarder/src/openapi.yaml
openapi: 3.0.3
info:
  title: WhatsApp Message Forwarding API
  description: |
    REST API for the WhatsApp Message Forwarding application.
    Receives incoming WhatsApp messages via Meta webhooks and forwards them to configured numbers.

    ## Authentication
    - **Admin endpoints** require an `adminToken` in the request body matching the `ADMIN_TOKEN` env variable.
    - **Webhook endpoints** are called by Meta and optionally verified via `X-Hub-Signature-256`.

    ## Frontend Testing
    Use this Swagger UI to test the `/health` and `/config/forward-number` endpoints directly.
    The `/webhook` endpoints are called by Meta — use the simulator examples below to test them locally.

  version: 1.0.0
  contact:
    name: mmtushar08
    url: https://github.com/mmtushar08/Whatsapp-Message-Forwarding-

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-ngrok-url.ngrok.io
    description: ngrok tunnel (for webhook testing)

tags:
  - name: Health
    description: Server health check
  - name: Webhook
    description: WhatsApp Cloud API webhook endpoints (called by Meta)
  - name: Config
    description: Runtime configuration management

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      description: Returns server status and timestamp. Use this to verify the server is running.
      operationId: healthCheck
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                    example: '2026-03-14T07:00:00.000Z'

  /webhook:
    get:
      tags: [Webhook]
      summary: Meta webhook verification handshake
      description: |
        Called by Meta when you register a webhook URL in the Meta Developer Dashboard.
        Meta sends hub.mode, hub.verify_token, and hub.challenge.
        If the token matches WEBHOOK_VERIFY_TOKEN, the server echoes back the challenge.
      operationId: verifyWebhook
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
            enum: [subscribe]
          example: subscribe
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
          example: my_secret_verify_token
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
          example: challenge_string_from_meta
      responses:
        '200':
          description: Verification successful — echoes back the challenge string
          content:
            text/plain:
              schema:
                type: string
                example: challenge_string_from_meta
        '403':
          description: Verification failed — token mismatch

    post:
      tags: [Webhook]
      summary: Receive incoming WhatsApp messages
      description: |
        Called by Meta when a WhatsApp message is received on your registered number.
        The server always responds 200 immediately (Meta requirement), then processes the message asynchronously.

        **Note:** When WHATSAPP_APP_SECRET is set, the X-Hub-Signature-256 header is verified.
      operationId: receiveWebhook
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: false
          schema:
            type: string
          description: HMAC-SHA256 signature of the request body (sent by Meta)
          example: sha256=abc123def456...
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
            example:
              object: whatsapp_business_account
              entry:
                - id: '123456789'
                  changes:
                    - value:
                        messaging_product: whatsapp
                        metadata:
                          display_phone_number: '15551234567'
                          phone_number_id: '987654321'
                        contacts:
                          - profile:
                              name: John Doe
                            wa_id: '15559876543'
                        messages:
                          - from: '15559876543'
                            id: wamid.abc123
                            timestamp: '1710400000'
                            type: text
                            text:
                              body: Hello, this is a test message!
                      field: messages
      responses:
        '200':
          description: Acknowledged (always returned immediately to Meta)
        '401':
          description: Invalid webhook signature (only when WHATSAPP_APP_SECRET is set)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /config/forward-number:
    patch:
      tags: [Config]
      summary: Update the forwarding phone number
      description: |
        Updates the phone number that messages are forwarded to at runtime.
        Requires a valid admin token matching the ADMIN_TOKEN environment variable.
        Changes take effect immediately without server restart.
      operationId: updateForwardNumber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateForwardNumberRequest'
            example:
              phoneNumber: '12345678900'
              adminToken: your_admin_token_here
      responses:
        '200':
          description: Phone number updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateForwardNumberResponse'
        '400':
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — invalid admin token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests (rate limited)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    WebhookPayload:
      type: object
      required: [object, entry]
      properties:
        object:
          type: string
          example: whatsapp_business_account
        entry:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              changes:
                type: array
                items:
                  type: object
                  properties:
                    value:
                      type: object
                    field:
                      type: string

    UpdateForwardNumberRequest:
      type: object
      required: [phoneNumber, adminToken]
      properties:
        phoneNumber:
          type: string
          description: Phone number with country code, digits only (no + sign)
          example: '12345678900'
          minLength: 7
          maxLength: 15
        adminToken:
          type: string
          description: Admin token matching ADMIN_TOKEN environment variable
          example: your_admin_token_here

    UpdateForwardNumberResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Forward-to number updated successfully
        previous:
          type: string
          example: '19876543210'
        current:
          type: string
          example: '12345678900'

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: 'Unauthorized: invalid admin token'